backend/requirements.txt
tensorflow==2.20.0
flask==3.1.0
flask-cors==5.0.0
Pillow==11.1.0
numpy==2.1.3
scikit-learn==1.6.1
matplotlib==3.10.1
seaborn==0.13.2
tqdm==4.67.1
opencv-python==4.10.0.84